+++ CHANGES.rst
@@ -1,6 +1,20 @@
 0.2 (unreleased)
 ----------------
 
+General
+^^^^^^^
+
+- Refactor sherpa models to use parent SherpaModelECPL class [#62]
+
+Bug Fixes
+^^^^^^^^^
+
+- Fix sherpa models guess() for integrated datasets.
+
+API Changes
+^^^^^^^^^^^
+
+- module sherpamod is now sherpa_modules.
 
 
 0.1 (2015-02-02)