+++ CHANGES.rst
@@ -6,7 +6,9 @@
   tables to SED or choose format of first data tables when providing multiple
   data tables.
 - Added support for  a ``flux_ul`` column in input data tables.
-- Added a method to estimate magnetic field: `naima.estimate_B`
+- Added a method to estimate magnetic field: `naima.estimate_B`.
+- Added the option to perform an optimization of the parameters before the MCMC
+  run: see option ``prefit`` in `naima.get_sampler`.
 
 Bug Fixes
 ^^^^^^^^^